adapt to test py

+++ app/services/url_service.py
@@ -3,10 +3,9 @@ import logging
 import re
 
 from models.url import Url
-from dbconfig import db, hashids
+from dbconfig import db
 from datetime import datetime, timedelta
-from werkzeug.exceptions import NotFound
-import base64
+from .id_hashing import INVALID_NUMBER, id_mapping
 
 def get_all_urls():
     '''
@@ -23,16 +22,17 @@ def create_new_url(body):
     :param body: request body
     :returns: the created entity
     '''
-    if (body is None or body.get('long_url') is None or is_valid_url(body.get('long_url'))) is False:
+    long_url = body.get('value')
+    if (long_url is None or is_valid_url(long_url)) is False:
         return None
     current_time = datetime.now()
-    existing_entity = Url.query.filter_by(long_url=body.get('long_url')).first()
+    existing_entity = Url.query.filter_by(long_url=long_url).first()
     if existing_entity is not None:
         if existing_entity.expire_date < current_time:
             ten_years_later = current_time + timedelta(days=365 * 10)
             existing_entity.expire_date = ten_years_later
             db.session.commit()
-        existing_entity.id = hashids.encode(existing_entity.id)
+        existing_entity.id = id_mapping.encode(existing_entity.id)
         return existing_entity.as_dict()
     expire_date = body.get('expire_date')
     if expire_date:
@@ -40,20 +40,23 @@ def create_new_url(body):
     else:
         expire_date = datetime.strptime('2029-12-31 23:59:59', '%Y-%m-%d %H:%M:%S')
 
-    url = Url(long_url=body.get('long_url'), expire_date=expire_date)
+    url = Url(long_url=long_url, expire_date=expire_date)
     db.session.add(url)
     db.session.commit()
-    url.id = hashids.encode(url.id)
+    url.id = id_mapping.encode(url.id)
     url = url.as_dict()
     return url
 
 def update_entity_by_identifier(identifier, body):
-    id = hashids.decode(identifier)
+    id = id_mapping.decode(identifier)
+    if id == INVALID_NUMBER:
+        return None
     entity = Url.query.filter_by(id=id).first()
     if entity is None:
         logging.info("identifier not found")
         return entity
-    if body is None or body.get('long_url') is None or is_valid_url(body.get('long_url')) is False:
+    long_url = body.get('url')
+    if long_url is None or is_valid_url(long_url) is False:
         raise Exception("illegal input")
     expire_date = body.get('expire_date')
     if expire_date:
@@ -63,14 +66,16 @@ def update_entity_by_identifier(identifier, body):
          expire_date = ten_years_later
 
     entity.expire_date = expire_date
-    entity.long_url = body.get('long_url')
+    entity.long_url = long_url
     db.session.commit()
-    entity.id = hashids.encode(entity.id)
+    entity.id = id_mapping.encode(entity.id)
     entity = entity.as_dict()
     return entity
 
 def get_url_by_identifier(identifier):
-    id = hashids.decode(identifier)
+    id = id_mapping.decode(identifier)
+    if id == INVALID_NUMBER:
+        return None
     current_time = datetime.now()
     entity = Url.query.filter_by(id = id).first()
     if entity is None:
@@ -81,7 +86,9 @@ def get_url_by_identifier(identifier):
     return entity.long_url
 
 def delete_by_identifier(identifier):
-    id = hashids.decode(identifier)
+    id = id_mapping.decode(identifier)
+    if id == INVALID_NUMBER:
+        return None
     entity = Url.query.filter_by(id = id).first()
     if entity:
         db.session.delete(entity)
@@ -98,48 +105,5 @@ def delete_all_urls():
 
 url_regex = re.compile(r'(?i)\b((?:[a-z][\w-]+:(?:/{1,3}|[a-z0-9%])|www\d{0,3}[.]|[a-z0-9.\-]+[.][a-z]{2,4}/)(?:[^\s()<>]+|\(([^\s()<>]+|(\([^\s()<>]+\)))*\))+(?:\(([^\s()<>]+|(\([^\s()<>]+\)))*\)|[^\s`!()\[\]{};:\'".,<>?«»“”‘’]))')
 
-
 def is_valid_url(url):
     return url is not None and bool(url_regex.search(url))
-
-# def get_url_by_long(body):
-#     # get the corresponding entity by short url, if not exist then return None
-#     # url = Url(**body)
-#     long_url = body['long_url']
-#     entity = Url.query.filter_by(long_url = long_url).first()
-    
-#     if entity:
-#         entity = entity.as_dict()
-#         print(type(entity))
-#         return entity
-#     else:
-#         return {'error': "404 not found"}
-
-# def update_shorturl(body):
-#     '''
-#     Update entity by id
-#     :param body: request body
-#     :returns: the updated entity
-#     '''
-#     user = User.query.get(body['id'])
-#     if user:
-#         user = User(**body)
-#         db.session.merge(user)
-#         db.session.flush()
-#         db.session.commit()
-#         return user
-#     raise NotFound('no such entity found with id=' + str(body['id']))
-
-# def delete_shorturl(id):
-#     '''
-#     Delete entity by id
-#     :param id: the entity id
-#     :returns: the response
-#     '''
-#     user = User.query.get(id)
-#     if user:
-#         db.session.delete(user)
-#         db.session.commit()
-#         return {'success': True}
-#     raise NotFound('no such entity found with id=' + str(id))
-
